refactor: use BigDecimal for amount parsing, document external_id scope
Two Minor findings from the final review, both cheap and worth taking:
Float for money going into a decimal(19,4) column is the classic "do not
use floats for money" smell even though it is currently exact for PYG
integers; BigDecimal matches how money is handled elsewhere in this app.
Also added a one-line comment on why external_id's hash doesn't include
account_id (the DB unique index already scopes by it, so this is
intentional, not an oversight for a future maintainer to "fix").

+++ app/models/android_purchase/webhook_processor.rb
@@ -61,11 +61,16 @@ class AndroidPurchase::WebhookProcessor
   private
 
     def numeric_amount
-      Float(@amount.to_s)
+      BigDecimal(@amount.to_s)
     rescue ArgumentError, TypeError
       nil
     end
 
+    # Deliberately excludes account_id: the DB unique index
+    # (index_entries_on_account_source_and_external_id) already scopes by
+    # account_id, so the same amount/timestamp/merchant hitting two
+    # different accounts (not currently possible -- one account per Tasker
+    # profile) would not collide across accounts.
     def external_id
       Digest::SHA256.hexdigest("#{@amount}|#{@timestamp}|#{@merchant}")
     end
